@@ -37,82 +37,88 @@
     "format": "npx prettier --write \"**/*.{js,mjs,cjs,ts,mts,cts,vue,json}\""
   },
   "devDependencies": {
-    "@constl/utils-tests": "^1.2.7",
-    "@eslint/js": "^9.7.0",
-    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@constl/utils-tests": "^1.6.7",
+    "@eslint/js": "^9.16.0",
+    "@mdi/font": "^7.4.47",
+    "@rollup/plugin-node-resolve": "^15.3.0",
     "@types/file-saver": "^2.0.7",
     "@types/js-cookie": "^3.0.6",
-    "@types/lodash": "^4.17.7",
-    "@types/node": "^18.19.39",
-    "@types/uuid": "^9.0.8",
+    "@types/lodash": "^4.17.13",
+    "@types/node": "^22.10.1",
+    "@types/uuid": "^10.0.0",
     "@types/webfontloader": "^1.6.38",
-    "@typescript-eslint/eslint-plugin": "7.13.1",
-    "@typescript-eslint/parser": "^7.16.1",
-    "@vitejs/plugin-vue": "5.0.5",
-    "@vue/test-utils": "2.4.6",
-    "cross-env": "7.0.3",
-    "electron": "31.0.2",
-    "electron-builder": "24.13.3",
-    "eslint": "9.5.0",
-    "eslint-plugin-vue": "9.26.0",
-    "happy-dom": "14.12.3",
-    "nano-staged": "0.8.0",
-    "node-pre-gyp": "^0.17.0",
-    "playwright": "1.44.1",
-    "prettier-plugin-organize-imports": "^4.0.0",
+    "@typescript-eslint/eslint-plugin": "8.14.0",
+    "@typescript-eslint/parser": "^8.16.0",
+    "@vitejs/plugin-vue": "^5.2.1",
+    "@vue/test-utils": "^2.4.6",
+    "cross-env": "^7.0.3",
+    "electron": "^33.2.1",
+    "electron-builder": "26.0.0-alpha.6",
+    "eslint": "^9.16.0",
+    "eslint-plugin-vue": "^9.32.0",
+    "happy-dom": "^14.12.3",
+    "nano-staged": "^0.8.0",
+    "playwright": "^1.49.0",
+    "prettier-plugin-organize-imports": "^4.1.0",
     "rollup-plugin-node-builtins": "^2.1.2",
     "rollup-plugin-node-globals": "^1.4.0",
     "rollup-plugin-resolve-id": "^1.0.2",
-    "simple-git-hooks": "2.11.1",
+    "sass": "^1.81.0",
+    "sass-loader": "^16.0.3",
+    "simple-git-hooks": "^2.11.1",
     "typed-emitter": "^2.1.0",
-    "typescript": "5.2.2",
-    "typescript-eslint": "^7.16.1",
+    "typescript": "5.6.3",
+    "typescript-eslint": "^8.16.0",
     "unplugin-auto-expose": "0.3.0",
-    "vite": "5.3.3",
+    "vite": "^6.0.3",
     "vite-plugin-node-polyfills": "^0.22.0",
-    "vite-plugin-vuetify": "^2.0.3",
-    "vitest": "2.0.3",
-    "vue-tsc": "2.0.26"
+    "vite-plugin-vuetify": "^2.0.4",
+    "vitest": "^2.1.6",
+    "vue-tsc": "2.1.10"
   },
   "dependencies": {
-    "@constl/ipa": "1.1.2",
-    "@constl/mandataire": "^2.0.8",
-    "@constl/mandataire-electron-principal": "^2.3.0",
-    "@constl/mandataire-electron-rendu": "^2.3.0",
-    "@constl/serveur": "^2.0.13",
-    "@constl/utils-ipa": "^1.0.3",
-    "@constl/vue": "^1.0.1",
+    "@constl/ipa": "1.5.1",
+    "@constl/mandataire": "^3.0.3",
+    "@constl/mandataire-electron-principal": "^3.0.4",
+    "@constl/mandataire-electron-rendu": "^3.0.4",
+    "@constl/serveur": "^2.1.4",
+    "@constl/utils-ipa": "^1.0.4",
+    "@constl/vue": "^1.0.4",
     "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
     "@ffmpeg/core": "^0.9.0",
     "@ffmpeg/ffmpeg": "^0.9.8",
-    "@mdi/font": "^7.4.47",
+    "@riffcc/orbiter": "^0.1.1",
+    "@vueuse/core": "^12.0.0",
     "ajv": "^8.17.1",
     "assert": "^2.1.0",
-    "core-js": "^3.37.1",
-    "crypto-browserify": "^3.12.0",
-    "electron-updater": "6.2.1",
+    "change-case": "^5.4.4",
+    "core-js": "^3.39.0",
+    "crypto-browserify": "^3.12.1",
+    "electron-updater": "6.3.9",
     "ennikkai": "^2.3.2",
     "events": "^3.3.0",
     "file-saver": "^2.0.5",
-    "hls.js": "^1.5.13",
+    "hls.js": "^1.5.17",
     "hlsjs-ipfs-loader": "^0.3.1",
     "image-file-resize": "^1.0.5",
-    "node-gyp-build": "^4.8.1",
+    "is-ipfs": "^8.0.4",
+    "node-gyp-build": "^4.8.4",
     "nuchabal": "^3.2.7",
     "semaphore-async-await": "^1.5.1",
     "tiny-typed-emitter": "^2.1.0",
-    "tslib": "^2.6.3",
-    "uuid": "^9.0.1",
-    "vue": "3.4.31",
-    "vue-router": "^4.4.0",
-    "vuetify": "^3.6.12",
+    "tslib": "^2.8.1",
+    "uuid": "^11.0.3",
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.0",
+    "vuetify": "^3.7.6",
     "webfontloader": "^1.6.28",
     "wherearewe": "^2.0.1"
   },
   "pnpm": {
     "overrides": {
-      "node-datachannel": "^0.9.0",
-      "@libp2p/autonat": "1.0.21"
+      "node-datachannel": "^0.20.0",
+      "@libp2p/autonat": "2.0.10",
+      "node-gyp": "^10.2.0"
     }
   }
 }
