@@ -1,21 +1,48 @@
 <template>
   <v-app>
-    <app-bar v-show="accountExists || enterAnonymously" />
-    <v-main>
-      <Home @enter="enterAnonymously = true" />
+    <app-bar />
+    <v-main min-height="100vh">
+      <router-view />
     </v-main>
+    <audio-player v-if="activeTrack"></audio-player>
+    <video-player
+      v-if="floatingVideoSource"
+      floating
+      :content-cid="floatingVideoSource"
+    ></video-player>
+    <app-footer />
   </v-app>
 </template>
 
 <script setup lang="ts">
-import {suivre as follow} from '@constl/vue';
-import {ref} from 'vue';
-import Home from './views/homePage.vue';
-import appBar from '/@/components/appBar.vue';
-import {useOrbiter} from '/@/plugins/orbiter/utils';
+import {onKeyStroke} from '@vueuse/core';
+import {ref, watchEffect} from 'vue';
+import audioPlayer from '/@/components/releases/audioPlayer.vue';
+import videoPlayer from '/@/components/releases/videoPlayer.vue';
 
-const {orbiter} = useOrbiter();
+import appBar from '/@/components/layout/appBar.vue';
+import appFooter from '/@/components/layout/appFooter.vue';
 
-const accountExists = follow(orbiter.listenForAccountExists.bind(orbiter));
-const enterAnonymously = ref(false);
+import {useAudioAlbum} from '/@/composables/audioAlbum';
+import {useFloatingVideo} from '/@/composables/floatingVideo';
+import {useShowDefederation} from '/@/composables/showDefed';
+
+const {showDefederation} = useShowDefederation();
+const {activeTrack} = useAudioAlbum();
+const {floatingVideoSource} = useFloatingVideo();
+
+const MAGIC_KEY = 'magicmagic';
+
+const yetToType = ref(MAGIC_KEY);
+onKeyStroke(e => {
+  if (!yetToType.value.length) return;
+  if (e.key === yetToType.value[0]) {
+    yetToType.value = yetToType.value.slice(1);
+  } else {
+    yetToType.value = MAGIC_KEY;
+  }
+});
+watchEffect(() => {
+  if (!yetToType.value.length) showDefederation.value = true;
+});
 </script>
