@@ -23,11 +23,11 @@
       </v-card-item>
       <v-card-text>
         <audio-viewer
-          v-if="release.release.type === 'audio'"
+          v-if="release.release.category === 'audio'"
           :file="release.release.file"
         />
         <video-viewer
-          v-else-if="release.release.type === 'movie'"
+          v-else-if="release.release.category === 'movie'"
           :file="release.release.file"
         />
       </v-card-text>
@@ -35,14 +35,14 @@
   </v-dialog>
 </template>
 <script setup lang="ts">
+import type {types as orbiterTypes} from '@riffcc/orbiter';
 import {ref} from 'vue';
-import type {ReleaseWithId} from '/@/plugins/orbiter/types';
 
 import AudioViewer from './audioViewer.vue';
 import VideoViewer from './videoViewer.vue';
 
 export interface ReleaseProps {
-  release: ReleaseWithId;
+  release: orbiterTypes.ReleaseWithId;
 }
 
 defineProps<ReleaseProps>();
