@@ -6,22 +6,51 @@
 
 // Styles
 import '@mdi/font/css/materialdesignicons.css';
-import 'vuetify/styles';
-
+import colors from 'vuetify/util/colors';
+import '/@/styles/main.scss';
+import '/@/styles/settings.scss';
 // Composables
 import {createVuetify} from 'vuetify';
-
+import {aliases, mdi} from 'vuetify/iconsets/mdi';
 // https://vuetifyjs.com/en/introduction/why-vuetify/#feature-guides
-export default createVuetify({
+const vuetify = createVuetify({
+  icons: {
+    defaultSet: 'mdi',
+    aliases,
+    sets: {
+      mdi,
+    },
+  },
+  defaults: {
+    global: {
+      elevation: 0,
+    },
+    VTextField: {
+      variant: 'solo-filled',
+    },
+    VFileInput: {
+      variant: 'solo-filled',
+    },
+    VSelect: {
+      variant: 'solo-filled',
+    },
+    VAutocomplete: {
+      variant: 'solo-filled',
+    },
+  },
   theme: {
     defaultTheme: 'dark',
     themes: {
-      light: {
+      dark: {
         colors: {
-          primary: '#1867C0',
-          secondary: '#5CBBF6',
+          primary: colors.purple.base,
+          'primary-lighten-1': colors.purple.lighten1,
+          'primary-darken-1': colors.purple.darken1,
+          'primary-accent': colors.purple.accent2,
         },
       },
     },
   },
 });
+
+export default vuetify;
