@@ -9,7 +9,7 @@ export async function loadFonts() {
 
   webFontLoader.load({
     google: {
-      families: ['Roboto:100,300,400,500,700,900&display=swap'],
+      families: ['Josefin Sans:100,300,400,500,700,900&display=swap'],
     },
   });
 }
