@@ -1,5 +1,5 @@
 <template>
-  <p></p>
+  <h1>Collection Page</h1>
 </template>
 <script setup lang="ts">
 defineProps<{collectionId: string}>();
