@@ -1,6 +1,6 @@
 declare module '*.vue' {
   import type {DefineComponent} from 'vue';
-  // eslint-disable-next-line @typescript-eslint/ban-types, @typescript-eslint/no-explicit-any
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-empty-object-type
   const component: DefineComponent<{}, {}, any>;
   export default component;
 }
